Added Assembler class as a place to start gluing together all the parsers and other bits.
